Fix RMSNormPerHead dtype: cast weight to input_dtype before multiplication
Prevents float32 output when input is bf16 (weight initialized as float32 zeros).

+++ dependency/verl_v070/verl/models/qwen3_6/megatron/layers/parallel_attention.py
@@ -125,7 +125,7 @@ class RMSNormPerHead(nn.Module):
     def forward(self, x):
         input_dtype = x.dtype
         output = self._norm(x.float()).to(input_dtype)
-        return output * self.weight
+        return output * self.weight.to(input_dtype)
 
 
 class ParallelQwen3_6Attention(nn.Module):
